+++ packages/kit/test/apps/async/src/routes/+page.svelte
@@ -5,7 +5,7 @@
 <h3>Tests</h3>
 <ul style:font-family="sans-serif">
 	{#each Object.keys(pages) as name (name)}
-		{@const href = name.slice(1, -13)}
+		{const href = name.slice(1, -13)}
 		<li>
 			<a {href}>{href}</a>
 		</li>