+++ packages/kit/test/apps/dev-only/src/hooks.client.js
@@ -2,7 +2,7 @@
 import cjs from 'e2e-test-dep-hooks-client';
 cjs.cjs();
 
-/** @type {import("@sveltejs/kit").HandleClientError} */
+/** @type {import("@sveltejs/kit/hooks").HandleClientError} */
 export function handleError({ kind, error }) {
 	if (kind !== 'unknown') return error;
 	return { message: /**@type{any}*/ (error).message };