+++ packages/kit/test/apps/options/source/hooks.server.js
@@ -4,7 +4,7 @@ import { building } from '$app/env';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { TOP_SECRET_SHH_PLS } from '$app/env/private';
 
-/** @type {import('@sveltejs/kit').Handle} */
+/** @type {import('@sveltejs/kit/hooks').Handle} */
 export function handle({ event, resolve }) {
 	return resolve(event, {
 		// this allows us to check that <link rel="stylesheet"> is still added