+++ packages/kit/test/types/app-error-enhanced/error.test.ts
@@ -1,4 +1,5 @@
-import { error, type HandleClientError, type HandleServerError } from '@sveltejs/kit';
+import { error } from '@sveltejs/kit';
+import { type HandleClientError, type HandleServerError } from '@sveltejs/kit/hooks';
 
 const app_error: App.Error = { status: 500, message: 'Unexpected error', additional: true };